....IJ.... [ZBX-5558,ZBX-5379] detecting a stale PID file
git-svn-id: svn://svn.zabbix.com/branches/dev@42887 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-5558/src/zabbix_java/shutdown.sh
@@ -5,11 +5,15 @@ source settings.sh
 
 if [ -n "$PID_FILE" ]; then
 	if [ -e "$PID_FILE" ]; then
-		kill `cat $PID_FILE`
-	else
-		echo "Zabbix Java Gateway is not running"
-		exit 1
+		PID=`cat "$PID_FILE"`
+		if ps -p "$PID" > /dev/null 2>&1; then
+			kill `cat $PID_FILE`
+			exit 0
+		fi
+		rm -f "$PID_FILE"
 	fi
+	echo "Zabbix Java Gateway is not running"
+	exit 1
 else
 	echo "Zabbix Java Gateway is not configured as a daemon: variable \$PID_FILE is not set"
 	exit 1
